Remove exception in ViewWrapperImpl::GetNextKeyboardFocusableActor
In NUI, it is not normal to create an empty handle.
Because all class instance is a reference type in C#, it is way natural to return null
instead an empty handle when the developer want to specify no actor can be focused
in GetNextKeyboardFocusableActor().

So the binding code should handle null for the return value of GetNextKeyboardFocusableActor(),
and not throwing an exception for this.

Change-Id: Ie12af44277743adea63bc45cfa8158bc6d4bce18
Signed-off-by: Jiyun Yang <ji.yang@samsung.com>