[AT-SPI] Exclude highlight frame from ShouldReportZeroChildren
When ShouldReportZeroChildren() returns true, any children of a given
actor are not displayed in the AT-SPI tree. This is good for composite
NUI controls (e.g. Button) so that they are displayed as a single entity
and not a relatively large subtree.

However, this introduces an inconsistency that the highlight frame actor
is sometimes also removed from the AT-SPI tree. This patch makes sure
the highlight actor is always reported.

Change-Id: I5e75af93289d9aca819a2cffeebd543a9270a9ca