Supports some tizen window features.
Supports some tizne window features.
1. Window resized and moved by disply server.
Supports window is resized or moved by display server.
This APIs should be called in mouse down event.
After that, the window is moved or resized with mouse move event.
When mouse up event happens, the job will be finished.

2. floating window mode.
The floating mode is to support making partial size window easliy.
It is useful to make popup style window and this window is always upper than the other normal window.
In addition, it is easy to change between popup style and normal style window.

Change-Id: I57f931325d6418fe157bb5eab5842e6e561bf2c9