Fix SuggestedMinimumWidth/Height to calculate required size correctly
SuggestedMinimumWidth/Height returns the minimum width and height of a
view required in a layout.

Previously, SuggestedMinimumWidth/Height used GetWidthForHeight and
GetHeightForWidth to calculate the minimum width and height of a view
although width/height were not dependent on height/width.
e.g.
if naturalSize.width = 0 and naturalSize.height = 100,
then SuggestedMinimumHeight would be 0 instead of 100.

Now, SuggestedMinimumWidth/Height uses GetWidthForHeight and
GetHeightforWidth only if width/height are dependent on height/width.

Change-Id: Ib65522b8e5399278b57a8866ad548e4d2be61469