[AT-SPI] Remove ShouldReportZeroChildren()
This method originated as a hack to cut the AT-SPI tree at a specific
node in the case of compound controls (i.e. composed of multiple views
overlaid on top of each other), so that there is only one leaf node
(instead of a subtree) for Button, Slider, etc.

However, the AccessibilityHidden property was introduced to allow for
controlling which nodes should be removed from the AT-SPI tree, and has
since become the recommended solution over ShouldReportZeroChildren(),
because the latter is less flexible ("all-or-nothing" approach) and
itself requires a hack in dali-csharp-binder to make sure the highlight
frame actor is never removed from the tree.

Change-Id: I761bfaf207e76a23962dcaa0b6c604a163cf8fe5