Add function to set maximum/minimum size
It is to set the maximized size when window is maximized or the size is increased by RequestResizeToServer.
Although the limitation is set by this function, window's size can be over the limitation by SetPositionSize() or SetSize().
The Set minimum size is similar to set maximum size.
Also it is only useful for tizen world.

Change-Id: I3e9724d0806f5b6cbfea7f0247272c4ab2472f5f